chore: refactor the `workers-shared` package and Wrangler builds
- The `workers-shared` package has been refactored to allow simpler imports
from other packages in the monorepo. It is no longer deployed to npm, as
it is not intended for public use.
- The Wrangler builds have been updated to replace the custom esbuild and `api-extractor`
with `tsup``, which simplifies the build setup and avoids issues with `api-extractor`.
- The package-lock file in import-npm fixture has been updated to use latest versions.
