+++ fixtures/workers-shared-asset-config/redirects.test.ts
@@ -1,14 +1,14 @@
+import Worker from "@cloudflare/workers-shared/asset-worker";
+import { normalizeConfiguration } from "@cloudflare/workers-shared/asset-worker/src/configuration";
+import { getAssetWithMetadataFromKV } from "@cloudflare/workers-shared/asset-worker/src/utils/kv";
 import { SELF } from "cloudflare:test";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { normalizeConfiguration } from "../../packages/workers-shared/asset-worker/src/configuration";
-import Worker from "../../packages/workers-shared/asset-worker/src/index";
-import { getAssetWithMetadataFromKV } from "../../packages/workers-shared/asset-worker/src/utils/kv";
-import type { AssetMetadata } from "../../packages/workers-shared/asset-worker/src/utils/kv";
+import type { AssetMetadata } from "@cloudflare/workers-shared/asset-worker/src/utils/kv";
 
 const IncomingRequest = Request<unknown, IncomingRequestCfProperties>;
 
-vi.mock("../../packages/workers-shared/asset-worker/src/utils/kv.ts");
-vi.mock("../../packages/workers-shared/asset-worker/src/configuration");
+vi.mock("@cloudflare/workers-shared/asset-worker/src/utils/kv.ts");
+vi.mock("@cloudflare/workers-shared/asset-worker/src/configuration");
 const existsMock = (fileList: Set<string>) => {
 	vi.spyOn(Worker.prototype, "unstable_exists").mockImplementation(
 		async (pathname: string) => {
@@ -40,8 +40,8 @@ describe("[Asset Worker] `test location rewrite`", () => {
 
 		const originalApplyConfigurationDefaults = (
 			await vi.importActual<
-				typeof import("../../packages/workers-shared/asset-worker/src/configuration")
-			>("../../packages/workers-shared/asset-worker/src/configuration")
+				typeof import("@cloudflare/workers-shared/asset-worker/src/configuration")
+			>("@cloudflare/workers-shared/asset-worker/src/configuration")
 		).normalizeConfiguration;
 		vi.mocked(normalizeConfiguration).mockImplementation(() => ({
 			...originalApplyConfigurationDefaults({}),
