+++ packages/miniflare/package.json
@@ -62,7 +62,7 @@
 		"@cloudflare/workers-shared": "workspace:*",
 		"@cloudflare/workers-types": "^4.20250525.0",
 		"@cloudflare/workflows-shared": "workspace:*",
-		"@microsoft/api-extractor": "7.49.1",
+		"@microsoft/api-extractor": "^7.52.8",
 		"@types/debug": "^4.1.7",
 		"@types/estree": "^1.0.0",
 		"@types/glob-to-regexp": "^0.4.1",
