+++ packages/miniflare/src/workers/assets/assets.worker.ts
@@ -1,5 +1,2 @@
-import worker from "@cloudflare/workers-shared/asset-worker/src/index";
-
-// Simply re-export the whole of the asset-worker so that it gets compiled into the Miniflare code base.
-// This allows us to have it as a devDependency only.
-export default worker;
+// Simply re-export the asset-worker so that its contents gets compiled into the Miniflare code base as a Worker template
+export { default } from "@cloudflare/workers-shared/asset-worker";
