+++ packages/vite-plugin-cloudflare/src/asset-workers/router-worker.ts
@@ -1,4 +1,1 @@
-// @ts-ignore
-import routerWorker from "@cloudflare/workers-shared/dist/router-worker.mjs";
-
-export default routerWorker;
+export { default } from "@cloudflare/workers-shared/router-worker";
