+++ packages/workers-shared/asset-worker/index.ts
@@ -1,0 +1,2 @@
+export * from "./src/worker";
+export { default } from "./src/worker";
