+++ packages/workers-shared/asset-worker/wrangler.jsonc
@@ -11,7 +11,7 @@
 	"name": "asset-worker",
 	"account_id": "0f1b8aa119a907021f659042f95ea9ba",
 	"workers_dev": false,
-	"main": "src/index.ts",
+	"main": "src/worker.ts",
 	"compatibility_date": "2024-07-31",
 	// nodejs_compat required when using @cloudflare/vitest-pool-workers
 	"compatibility_flags": ["nodejs_compat"],
