+++ packages/workers-shared/index.ts
@@ -1,3 +1,6 @@
 export * from "./utils/constants";
 export * from "./utils/types";
 export * from "./utils/helpers";
+export * from "./utils/configuration/parseHeaders";
+export * from "./utils/configuration/parseRedirects";
+export * from "./utils/configuration/types";
