+++ packages/workers-shared/turbo.json
@@ -1,9 +1,5 @@
 {
 	"$schema": "http://turbo.build/schema.json",
 	"extends": ["//"],
-	"tasks": {
-		"build": {
-			"outputs": ["dist/**"]
-		}
-	}
+	"tasks": {}
 }
