+++ packages/workers-shared/utils/tsconfig.json
@@ -11,6 +11,7 @@
 		"forceConsistentCasingInFileNames": true,
 		"strict": true,
 		"noUncheckedIndexedAccess": true,
+		"noImplicitOverride": true,
 		"skipLibCheck": true
 	}
 }
