+++ packages/wrangler/src/__tests__/deployments.test.ts
@@ -1,5 +1,5 @@
 import fs from "node:fs";
-import { thrownIsDoesNotExistError } from "../utils/filesystem";
+import { thrownIsDoesNotExistError } from "@cloudflare/workers-shared";
 import { mockAccountId, mockApiToken } from "./helpers/mock-account-id";
 import { mockConsoleMethods } from "./helpers/mock-console";
 import { clearDialogs } from "./helpers/mock-dialogs";
